requirements.txt
aiodns==1.1.1
aiohttp==2.0.7
async-timeout==1.2.0
cchardet==2.0.0
cffi==1.10.0
chardet==3.0.2
discord.py==0.16.7
multidict==2.1.4
overwatch-api==0.5
packaging==16.8
pycares==2.1.1
pycparser==2.17
PyNaCl==1.1.2
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
websockets==3.3
yarl==0.10.0
youtube-dl==2017.4.28
python-Levenshtein==0.12.0